Simplify code to have an off-canvas menu
While Foundations's off-canvas menu allows us to forget about writing
CSS, it also leads to complicated HTML.

Ideally Foundation would provide an easy way to simplify what we're
doing, but I haven't found anything in the documentation.

We could simplify the HTML a bit more if we used a CSS grid layout
instead of a flex one, but old browsers have better support for the
latter.

Note we're using `breakpoint(medium)` so we can group the CSS for small
screens and follow SCSS-Lint rules at the same time.

Also note behavior of the main area when the menu appears on small
screens is slightly different: instead of moving its content to the
right, it becomes smaller. I've done it this way so we don't have any
overflow issues, unlike the previous version.

There's a small issue using a label and a checkbox to enable/disable the
menu: sighted keyboard users with a small screen will never be able to
enable the menu. We believe sighted keyboard users don't use small
screens, particularly in the admin area, and so they won't be affected
by this change. But if you think we're wrong, please let us know ;). On
the plus side, using a checkbox works without JavaScript.